temporary fix for IE css loading
the client renderer wasn't waiting for css files to load if it found a
link tag with the same path already present in the head, but this didn't
mean the actual response for the css had arrived. this is now fixed, but
image requests are made every time a css is requested, even though
previous image requests have been done for the same file. moreover, this
is a temporary fix, since css dependecy management has to be done on the
server, since IE8 & 9 have a hard limit of 31 style sheets on one page
and this can be avoided with correct aggregation on the server only.